Fixed Many bugs - Canvases only work in Screen space as a quick fix for performance     - this will be fixed when ever required for a game I am working on - Updated all size_t types that are not for sizes to be uint64_t - Removed the Object "createDestroy()" function and replaced it with a virtual destructor     - Removed all implementations of the createDestroy function from objects already implemented - Fixed bug with max interpolation variable - Created Setting, Settings, SettingsUI classes - Fixed bug with colliders not being destroyed properly - Started working on a improved particle system

+++ include/ObjectManager.hpp
@@ -11,9 +11,9 @@ class ObjectManager
 {
 public:
     /// @returns nullptr if the object does not exist 
-    static Object::Ptr<> getObject(size_t id);
+    static Object::Ptr<> getObject(uint64_t id);
 
-    static size_t getNumberOfObjects();
+    static uint64_t getNumberOfObjects();
 
     /// @brief destroys all the objects in the queue
     /// @note only use this if you know what you are doing
@@ -38,8 +38,8 @@ private:
     class _objectCompClass : public Object
     {
     public:
-        _objectCompClass(size_t id);
-        void setID(size_t id);
+        _objectCompClass(uint64_t id);
+        void setID(uint64_t id);
         createDestroy();
     };
